[WIP] Fix analyzer violations in BotCore Services (#480)
* Initial plan

* fix(services): Batch 1 - Performance and async improvements (18 violations fixed)


* fix(services): Batch 2 - Removed NullReferenceException catches and nested ternary (36 violations fixed)


* Changes before error encountered


* Changes before error encountered


* fix(services): Batch 3 - Remove unnecessary casts (4 violations fixed)


* fix(services): Batch 4 - Simplify LINQ expressions (4 violations fixed)


* fix(services): Batch 5 - Identifier naming compliance (6 violations fixed)


---------

+++ src/BotCore/Services/MarketConditionAnalyzer.cs
@@ -296,7 +296,7 @@ public class MarketConditionAnalyzer
             return new TrendAnalysis
             {
                 Direction = direction,
-                Strength = Math.Min(1m, (decimal)trendStrength * TrendStrengthScalingFactor), // Scale to 0-1
+                Strength = Math.Min(1m, trendStrength * TrendStrengthScalingFactor), // Scale to 0-1
                 ShortTermMA = shortMA,
                 MediumTermMA = mediumMA,
                 LongTermMA = longMA
